Started rewrite of sources.py, beginning of fringeSelect which can help you find a fringe finder
